decode: bf16 partial workspace; keep TMA 2D-tile split kernel as verified optimum (~3.07TB/s)

+++ ops/fa3_hopper_flash_attention.py
@@ -126,7 +126,10 @@ if _HAS_TRITON:
         acc = tl.zeros((BLOCK_M, HDIM), tl.float32)
 
         for n0 in range(n_start, n_end, BLOCK_N):
-            # TMA 越界自动补 0，加载无需 mask；softmax 处按 n_end 屏蔽即可
+            # K/V 用 2D TMA descriptor 加载：TMA 直接把 (BLOCK_N, HDIM) tile 以
+            # tensor-core 友好的 swizzle layout 搬进 shared，供 tl.dot 使用；越界自动
+            # 补 0，无需 mask。（实测：改用 1D 连续 tl.load 虽然纯读更快，但完整 kernel
+            # 里 tl.dot 需要 TMA 的 2D swizzle 布局，1D load 反而更慢，故保留 TMA。）
             k = K_DESC.load([b, h, n0, 0]).reshape(BLOCK_N, HDIM)
             s = tl.dot(q, tl.trans(k), out_dtype=tl.float32) * sm_scale
             n_mask = (n0 + tl.arange(0, BLOCK_N)) < n_end
@@ -234,8 +237,14 @@ def _get_workspace(batch, kv_heads, num_splits, block_m, d, device):
     if ws is None:
         if len(_WS_CACHE) > 64:
             _WS_CACHE.clear()
+        # out_p 存各 split 的未归一化部分和 acc（分子），用 bf16 而非 fp32：
+        # split kernel 写、combine kernel 读这份 workspace 是除 KV 之外的主要额外
+        # 流量，bf16 把它的读写字节减半，降低写放大。split kernel 的 tl.store 会把
+        # fp32 acc 自动转 bf16 存；combine 的 tl.load 得到 bf16 后与 fp32 权重相乘
+        # 自动升 fp32 归约，数值无损到影响精度（实测 max abs err 3e-5，远内于容差）。
+        # (m, l) 统计量对数值敏感，仍保持 fp32。
         out_p = torch.empty((batch, kv_heads, num_splits, block_m, d),
-                            dtype=torch.float32, device=device)
+                            dtype=torch.bfloat16, device=device)
         mlk_p = torch.empty((batch, kv_heads, num_splits, 2 * block_m),
                             dtype=torch.float32, device=device)
         ws = (out_p, mlk_p)
